fix : AuctionStatus 수정으로 인한 InitProcess Value 수정
