+++ src/main/java/com/deal4u/fourplease/domain/member/mypage/dto/MyPageAuctionHistory.java
@@ -1,0 +1,4 @@
+package com.deal4u.fourplease.domain.member.mypage.dto;
+
+public record MyPageAuctionHistory() {
+}
