+++ src/main/java/com/deal4u/fourplease/domain/member/mypage/service/MyPageAuctionHistoryService.java
@@ -1,0 +1,4 @@
+package com.deal4u.fourplease.domain.member.mypage.service;
+
+public class MyPageAuctionHistoryService {
+}
